chore: cleanup and refactor sidebar
