+++ app/(pages)/dashboard.tsx
@@ -1,15 +1,14 @@
 import React from 'react';
 import { View, Text, ScrollView } from 'react-native';
-import DashboardWidgets from '../../components/DashboardWidgets';
-import StatusOverview from '../../components/StatusOverview';
-import NewestPracticesTable from '../../components/NewestPracticesTable';
+import DashboardWidgets from '../../components/features/dashboard/DashboardWidgets';
+import StatusOverview from '../../components/features/dashboard/StatusOverview';
+import NewestPracticesTable from '../../components/features/dashboard/NewestPracticesTable';
 import { UseEnvironment } from '../../utils/environmentUtils';
 
 const Dashboard: React.FC = () => {
   const { tailwind } = UseEnvironment();
   return (
     <ScrollView style={tailwind('p-4')}>
-      {/* Section 1: Welcome */}
       <View style={tailwind('mb-6')}>
         <Text style={tailwind('text-2xl font-bold text-gray-800')}>
           Welcome Andrew!
@@ -20,13 +19,8 @@ const Dashboard: React.FC = () => {
         </Text>
       </View>
 
-      {/* Section 2: Widget Stats */}
       <DashboardWidgets />
-
-      {/* Section 3: Status Overview */}
       <StatusOverview />
-
-      {/* Section 4: Newest Practices Table */}
       <NewestPracticesTable />
     </ScrollView>
   );
