+++ app/_layout.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Slot } from 'expo-router';
 import { DropdownProvider } from '../context/DropdownProvider';
-import Layout from '../components/Layout';
+import Layout from '../components/layout/Layout';
 
 export default function RootLayout() {
   return (
