+++ components/features/dashboard/CardView.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { View, Text, TouchableOpacity, Image, Switch } from 'react-native';
-import PencilIcon from '../../assets/images/pencil.svg';
-import DeleteIcon from '../../assets/images/trash.svg';
-import { UseEnvironment } from '../../utils/environmentUtils';
+import PencilIcon from '../../../assets/images/pencil.svg';
+import DeleteIcon from '../../../assets/images/trash.svg';
+import { UseEnvironment } from '../../../utils/environmentUtils';
 
 interface CardProps {
   row: {
@@ -19,16 +19,14 @@ export const CardView: React.FC<CardProps> = ({ row, toggleStatus }) => {
   const { tailwind } = UseEnvironment();
 
   return (
-    <View style={tailwind('p-4 mb-4 rounded-lg p-4 ')}>
+    <View style={tailwind('p-4 mb-4 rounded-lg p-4')}>
       <Text style={tailwind('text-lg font-bold mb-2')}>{row.name}</Text>
       <Text style={tailwind('text-sm text-gray-600')}>{row.phone}</Text>
       <Text style={tailwind('text-sm text-gray-600')}>{row.email}</Text>
       <Text style={tailwind('text-sm text-gray-600 mb-2')}>
         Created: {row.date}
       </Text>
-
       <View style={tailwind('flex-row items-center justify-between')}>
-        {/* Status Toggle */}
         <View style={tailwind('flex-row items-center')}>
           <Switch
             value={row.active}
@@ -41,19 +39,18 @@ export const CardView: React.FC<CardProps> = ({ row, toggleStatus }) => {
           </Text>
         </View>
 
-        {/* Actions */}
         <View style={tailwind('flex-row')}>
           <TouchableOpacity style={tailwind('mr-3')}>
             <Image
               source={PencilIcon as any}
-              style={{ width: 22, height: 22 }}
+              style={{ width: 14, height: 14 }}
               tintColor="#67adb9"
             />
           </TouchableOpacity>
           <TouchableOpacity>
             <Image
               source={DeleteIcon as any}
-              style={{ width: 22, height: 22 }}
+              style={{ width: 14, height: 14 }}
               tintColor="#67adb9"
             />
           </TouchableOpacity>
