+++ components/layout/Sidebar.tsx
@@ -1,50 +1,17 @@
 import React from 'react';
-import { View, Text, Image } from 'react-native';
-import { useRouter, usePathname } from 'expo-router';
-import { UseEnvironment } from '../../utils/environmentUtils';
-import SidebarItem from './SidebarItem';
+import { View } from 'react-native';
+import { useSidebar } from './SidebarContext';
+import SidebarMenu from './SidebarMenu';
 import SidebarLogo from './SidebarLogo';
-import { MenuItem, menuItemsData } from './Sidebar.data';
-import { getCustomStyles } from './styles';
-import { useDropdown } from '../../context/DropdownProvider';
-
+import { UseEnvironment } from '../../utils/environmentUtils';
 import closeSquare from '../../assets/images/close-square.svg';
 import caretRight from '../../assets/images/caret-right.svg';
-import MedicalIcon from '../../assets/images/medical-briefcase.svg';
+import SidebarItem from './SidebarItem';
 
 const Sidebar: React.FC = () => {
-  const { environment, tailwind, breakpoints } = UseEnvironment();
-  const router = useRouter();
-  const pathname = usePathname();
-  const { activeDropdown } = useDropdown();
-  const [isCollapsed, setIsCollapsed] = React.useState(true);
+  const { isCollapsed, toggleSidebar } = useSidebar();
+  const { environment, breakpoints, tailwind } = UseEnvironment();
   const { width } = environment;
-  const customIconColor =
-    (tailwind('text-customIcon')?.color as string) || '#67adb9';
-
-  const toggleSidebar = () => setIsCollapsed(!isCollapsed);
-  const isNotificationsOpen = activeDropdown === 'notifications';
-  const isProfileOpen = activeDropdown === 'profile';
-
-  const sidebarItems: MenuItem[] = React.useMemo(() => {
-    return menuItemsData.map((item) => {
-      return {
-        ...item,
-        isVisible:
-          isProfileOpen || isNotificationsOpen
-            ? [
-                'Dashboard',
-                'My Profile',
-                'Manage Users',
-                'Manage Patients',
-                'Logs',
-              ].includes(item.name)
-            : ['Dashboard', 'My Profile', 'Manage Practices', 'Logs'].includes(
-                item.name
-              ),
-      };
-    });
-  }, [isProfileOpen, isNotificationsOpen]);
 
   return (
     <View
@@ -78,7 +45,9 @@ const Sidebar: React.FC = () => {
             onPress={toggleSidebar}
             backgroundColor="transparent"
             textColor="black"
-            iconColor={customIconColor}
+            iconColor={
+              (tailwind('text-customIcon')?.color as string) || '#67adb9'
+            }
             isCollapsed={false}
           />
         </View>
@@ -86,63 +55,7 @@ const Sidebar: React.FC = () => {
         <SidebarLogo />
       )}
 
-      <View
-        style={tailwind(
-          isCollapsed && width < breakpoints.md ? 'px-5' : 'px-15'
-        )}
-      >
-        {isNotificationsOpen && (
-          <View
-            style={[
-              tailwind('flex-row items-center rounded-lg mb-3'),
-              {
-                backgroundColor: '#f8f8f8',
-                paddingVertical: 18,
-                paddingHorizontal:
-                  isCollapsed && width < breakpoints.md ? 5 : 12,
-              },
-            ]}
-          >
-            <View style={tailwind('w-30 items-center')}>
-              <Image
-                source={MedicalIcon as any}
-                style={{ width: 22, height: 22 }}
-                tintColor="#578388"
-              />
-            </View>
-            <Text
-              style={[
-                tailwind('font-bold ml-4'),
-                {
-                  fontSize: 18,
-                  color: '#333',
-                  opacity: width >= breakpoints.md || !isCollapsed ? 1 : 0,
-                },
-              ]}
-            >
-              Cape Fertility Clinic
-            </Text>
-          </View>
-        )}
-
-        {sidebarItems.map((item: MenuItem) => {
-          const isActive = pathname === item.path;
-          const styles = getCustomStyles(isActive);
-          if (!item.isVisible) return null;
-
-          return (
-            <SidebarItem
-              key={item.path}
-              item={item}
-              onPress={() => router.push(item.path as any)}
-              backgroundColor={styles.backgroundColor}
-              textColor={styles.color ?? 'black'}
-              iconColor={styles.iconColor ?? 'gray'}
-              isCollapsed={isCollapsed && width < breakpoints.md}
-            />
-          );
-        })}
-      </View>
+      <SidebarMenu isCollapsed={isCollapsed} />
     </View>
   );
 };
