+++ components/layout/SidebarContext.tsx
@@ -1,0 +1,34 @@
+import React, { createContext, ReactNode, useContext, useState } from 'react';
+
+interface SidebarContextProps {
+  isCollapsed: boolean;
+  toggleSidebar: () => void;
+}
+interface SidebarProviderProps {
+  children: ReactNode;
+}
+
+const SidebarContext = createContext<SidebarContextProps | undefined>(
+  undefined
+);
+
+export const SidebarProvider: React.FC<SidebarProviderProps> = ({
+  children,
+}) => {
+  const [isCollapsed, setIsCollapsed] = useState(true);
+  const toggleSidebar = () => setIsCollapsed(!isCollapsed);
+
+  return (
+    <SidebarContext.Provider value={{ isCollapsed, toggleSidebar }}>
+      {children}
+    </SidebarContext.Provider>
+  );
+};
+
+export const useSidebar = () => {
+  const context = useContext(SidebarContext);
+  if (context === undefined) {
+    throw new Error('useSidebar must be used within a SidebarProvider');
+  }
+  return context;
+};
